requirements.txt
azure.cognitiveservices.speech
google.generativeai
google.cloud
pywhatkit
spotipy
requests